@@ -4,15 +4,15 @@ description: Knowles' Audible Navigation for Visually Impaired Sailors
 version: 1.0.0+4
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_tts:
+  flutter_tts: ^3.2.4
   intl:
   shared_preferences:
-  audioplayers: ^0.14.0
+  audioplayers: ^0.20.1
   nmea:
     git:
       url: https://github.com/jamesdalby/nmea.git
@@ -20,12 +20,13 @@ dependencies:
     git:
       url: https://github.com/jamesdalby/ais.git
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.4
+  collection: ^1.15.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: "^0.7.3"
+  flutter_launcher_icons: ^0.9.2
 
 flutter_icons:
   android: true
@@ -37,12 +38,4 @@ flutter:
   uses-material-design: true
 
   assets:
-    - assets/beeps/medium-1.mp3
-    - assets/beeps/high-1.mp3
-    - assets/beeps/high-2.mp3
-    - assets/beeps/high-3.mp3
-    - assets/beeps/low-1.mp3
-    - assets/beeps/low-2.mp3
-    - assets/beeps/low-3.mp3
-    - assets/beeps/upchirp.mp3
-    - assets/beeps/downchirp.mp3
+    - assets/beeps/
